[taskscheduler] Add stackdriver tracing
Before this lands, the relevant service task scheduler service
accounts will need to be granted the "Cloud Trace Agent"
role.

This plumbs context through a lot of places and adds a lot of
trace spans centered around the MainLoop and various other
functions that I saw that looked interesting.

The goal is to make it easier to determine bottlenecks in
taskdriver. In Perf and Gold, we've found the StackDriver
tracing to be very useful for this.

The span naming convention is the same one I use in Gold -
for private functions/methods, it's just the name of the
method. For public functions/methods, it has a prefix that
indicates which file it is in. This makes it a bit easier
to track through the code (since the spans don't know what
file number or such they are in by default).

Change-Id: Ie5d2c3c84aa0e3a3c08008a5ae29e98a96bd7098
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/452745
Auto-Submit: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Eric Boren <borenet@google.com>