[skcq] Move AUTHORS verification directly into SkCQ
The AUTHORS file check in PRESUBMIT.py only checks the owner of the CL
and not the actual Gerrit author of the CL.
Doing it in PRESUBMIT.py with python is harder to maintain and test.
SkCQ seems like a better place because of the verifications oriented
approach it takes.

Bug: skia:12487
Change-Id: Ib17771b20b5116334d06356851d9c7d4c2bef002
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/452350
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Ravi Mistry <rmistry@google.com>