[infra] Update RBE image to be based off Debian10.
The old RBE image was based off of Ubunutu 16.04 Xenial
and this was problematic when running our Webpack-based
Puppeteer tests. The way we support those is by installing
node and a few other deps on top of the RBE image and then
running the tests. The Node 14 installer stopped working
for that version of Xenial and running apt-get upgrade took
way way too long to be feasible.

We are likely to run into issues like this as a result of
using such an old image as our base more and more frequently.
When we work on the Skia RBE image, we'll want more control
over our C++ toolchain than what happens to be pre-installed.

Both of those are the impetus to base our image of off something
newer and control more of the packages. Debian 10 was chosen
because Debian 11 is not quite available from the Cloud image
store (yet, see https://github.com/GoogleContainerTools/base-images-docker/tree/master/debian11).

While putting this together, I learned a few facts and
techniques:
  1) The rbe_images served by Google Cloud are configured/setup
     in the layer-definitions GitHub repo (example: https://github.com/GoogleCloudPlatform/layer-definitions/blob/ef7ab83fe8f7813ceeef5bb8dc283a08bed47151/layers/ubuntu1604/rbe_tools/BUILD)
  2) We can use http_archive to download an SDK or similar from
     a trusted source, verify the sha256 hash, and then extract
     and install it using pkg_tar and language_tool_layer.
     See patchset 8 where this works for the Android NDK also.
     The infra team doesn't need the NDK anymore, so it was removed
     in a later patchset.
  3) We can use language_tool_layer and toolchain_container
     to better organize the creation of our RBE docker image.
     This takes our previous approach (akin to implementing
     everything in main()) and breaks it into pieces.
  4) We *do* need the Java toolchain to run certain Bazel commands.
  5) Some of our tests secretly depended on Python2, which is not
     installed in the new image. Those have been updated.

Suggested Review Order:
 - bazel/rbe/README.md (Reminder that the rest of bazel/rbe is
   generated)
 - BUILD.bazel and WORKSPACE
 - go code for test fixes
 - everything else

BYPASS_INCLUSIVE_LANGUAGE_REASON=generated_code
Change-Id: I3a92fcae6cfc9334e4fe943d632db67c77b6d26f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/455136
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>