[gold] Add in RPC for Chrome's gold_inexact_matching tool.
The /json/v1/positivedigestsbygrouping/ RPC adds a way to
get all positively triaged images for the given traces
in a specified grouping (aka test).

Since we are using groupingID now as part of a public API,
I added some constants and tests to verify that we don't
accidentally change how groupingID is computed (and that
our test data conforms to that calculation).

Change-Id: Ibe7a9dd5c029049441add3d8bbfc5c6861762fda
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/456486
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>