[ansible] Deploy metadata_server_ansible via CIPD.
This adds the build_and_release_metadata_server_ansible.yml playbook
which is used to build and release the binary, which is required
since the binary is built with the service account key, and
we entrust that secret handling to Ansible.

Note that the Makefile targets added to build metadata_server_ansible
check for empty keys in the environment variables, which prevents
pushing versions of the metadata server that don't have keys
embedded.

Also includes a drive-by cleanup of upgrade-win-machines-to-python3.yml.

Bugs: skia:12448
Change-Id: I8ebce1d25e8fbfdc87dcc453de1c8a9529c14535
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/458440
Reviewed-by: Erik Rose <erikrose@google.com>