[task scheduler] Explore adding focused, concise tests.
It is recommended to look at this CL patchset by patchset because
a lot is happening all at once.

PS 1 is https://skia-review.googlesource.com/c/buildbot/+/451836

PS 2 adds some focused tests to scoreCandidate for tryjob and
     forced jobs (since those were easy for me to understand).

PS 3 tweaks ComputeBlamelist a bit to make the parameters fake-able
     and adds a few test utils for setting up a DAG of commits,
     fake tests, fake windows, etc. Then it creates a bunch of
     of tests for ComputeBlamelist.


The goal is that these smaller tests are easier to understand
and reason about - the setup is contained and clear, the
assertions are direct.

Change-Id: I3779e398d4b2a592608898f29bfdf7a859b7b54a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/457092
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>