[gold] Move cros-tast-dev to https://cros-tast-gold.skia.org/
With this configuration, we are ingesting data from
gs://skia-gold-cros-tast-dev
and
gs://skia-gold-cros-tast

I've kept the internal instance name the same, as that is less risky
than copying all the data over. The only data I needed to copy
from the old bucket to the new buckets were the images, which I did
with gsutil.

Change-Id: I98308a54079c4552b55472c0f12b0549f1dc0cfa
Bug: skia:10808
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/459440
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>