[machine] Fix battery level reading on Hammerhead devices.
Change the dumpsys regexs to use \b instead of \n to detect the end
of the match.

Hammerhead devices (or at least the version of Android we run
on them) returns dumpsys lines separated with \r\n instead of
the \n we were searching for.

Adds test case and more logging to help debug future errors.

Change-Id: Ieb2e21acb7e5de5ce5ecf45a39fd5f3d4fb8626f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/460876
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>