[gold] Handle duplicated patchset ids belonging to different CLs
Occasionally, when ingesting GitHub data, we see the same commit
uploaded to two different PRs (CLs). This causes data from the
second PR to be uploaded as belonging to the first PR because the
Patchset already exists with the primary key belonging to that
specific commit.

This does not happen in practice with Gerrit because Gerrit includes
a Change-Id that is random each time, so the likelihood of a
collision is much lower.

Regardless, if we detect a duplicate, we append the CL id onto
the end of the primary key for the patchset. This place, in
tryjob_ingestion, is the only place that cares about calling
a Patchset something consistent, so we don't need to update
any other callsites.

I choose not to add this suffix for all patchsets because
I didn't want to mess up any data that has been ingested
to a patchset with one name, and then gets split into
a different PS after this lands. To properly handle that would
require a lot of data migration in the tables.

Change-Id: Ia93f56e711b0dae4506804da92f952b076fae339
Bug: skia:12580
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/463434
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>