[infra] Port jsfiddle to Bazel
The webpack version appears to still build, so landing this
shouldn't break the existing auto-deploy flow. If it does,
we'll roll back the auto-build and try to land the Bazel
version ASAP.

To make our gazelle extension happy, I needed to rename
the elements to have a -sk suffix.

Use of @use tags is preferred because they turn into a
SCSS compile error if the files are not found.
@import is apparently deprecated https://sass-lang.com/documentation/at-rules/import

To get the CanvasKit and PathKit binaries into the build,
I made jsfiddle/build/BUILD.bazel which refers to the files,
and checked in placeholders. One can run
`make wasm_libs` from //jsfiddle to download them to the
correct place. This should be replaced with a genrule
when we do not need the webpack build. This should be
flexible enough to work with the autodeploy script,
which can copy the ToT binaries into jsfiddle/build
before running `make bazel_release`

This also adds copy_files to our sk_page macro as a
way to get files into the final build. This allows us
to get in the JS and WASM parts of CanvasKit/PathKit
without having to import them in a way that makes
WebPack or Bazel sad.

Change-Id: I6a8b5790b45a4cbe684457f57bd95982d73f77d7
Bug: skia:10614
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/479051
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>