[infra] Update autodeploy logic to work with Bazel jsfiddle
This works differently than our existing PushAppsFromWASMDockerImage
task. The previous way uses the latest built docker image
containing the skia infra checkout and a volume containing the
built JS+WASM blobs from the specified skia revision to build the
server binaries and put those into a docker container that will
be ran on k8s.

This approach checks out the skia infra repo at a configurable
revision (defaults to tip of tree) and runs (effectively)
bazel build from there (I didn't want to run bazel in docker
as that seemed difficult and unnecessarily complex). The
bazel build uses the JS+WASM binaries from the specified
skia revision.

This could be made to work with an image from a tryjob
if we specify the skia_revision as "[issue]_[patchset]",
which could be handled via gen_tasks_logic and not here.

This does not yet publish to PubSub, but that is the next
step (along with removing the old way from
push_apps_from_wasm_image.go)

Change-Id: I02cd92115c2782ae487e4a9a36e6428104b2327e
Bug: skia:10614
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/481141
Reviewed-by: Eric Boren <borenet@google.com>