[shaders] Port to Bazel
This adds some arguments to the karma_test rule which allow us
to serve static files alongside the test as well as provide a
custom karma.conf.js.

This removes the checked in images (according to .gitignore,
it was intended for them to be ignored?) and uses a genrule to
get them for bundling into the package.

This also fixes and works around a few new TS warnings.

Change-Id: Ie1a4909314098c215804d0ba2c2a8301ba706db0
Bug: skia:10614
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/486659
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>