[bazel] Simplify test_on_env and make compatible with Mac
getopt is not cross-platform and was not working on Macs.
I tried to use getopts (notice the s) that is cross-platform
(a built in bash function), but found it to be fairly complicated.

Because this is an internal script that is called in once place
by generated Bazel rules, I removed the named arguments and went
with three required fixed-location arguments. This should make
the script easier to read.

Change-Id: I7cfb968496d3404a71a9999aeb3db412c8d69dc7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/491049
Reviewed-by: John Stiles <johnstiles@google.com>