Add breakpoints to SkSL Debugger.
A subtle ◯ marker in the gutter/line-number region of the code editor
indicates that a breakpoint can be set on a line. Clicking the ◯ marker
will set an actual breakpoint, indicated with a less-subtle ● marker.

A new Run/Run to Breakpoint button has been added to execute the shader
until a breakpoint is hit or execution completes.

Change-Id: I2e5228f6044517a1d05668d87f128424a86321a8
Bug: skia:12778
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/492596
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>