Ensure that the breakpoint mark doesn't change size.
The characters ◯ and ● aren't necessarily the same size and position on
every platform. To guarantee the bullet always looks the same, we now
use ● all the time, and give it a tiny red shadow. To show only an
outline, the character color is changed to match the background color,
so only the red shadow is visible. To show a filled-in breakpoint, the
character color is changed to red.

Bug: skia:12778
Change-Id: I1cb6b69c232860da5d000497feb7f37d9f92bd74
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/492676
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>