Enable basic SkSL debugging on shaders.skia.org.
The Shaders page now has a Debug button which generates a debug trace
from the shader. It's currently hard-coded to sample from the center of
the viewport. Once the debug trace has been made, it's copied to Local
Storage and the SkSL Debugger is opened in a new tab.

The followup CL allows the user to set the debug trace coordinate via
clicking directly in the canvas.

Bug: skia:12818
Change-Id: I6d32d923171e17a976b53fdd3f41e00b800a9178
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/495423
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>