[skolo] Install UltraVNC server on all Windows machines.
TightVNC produces a black screen on some Windows machines. With UltraVNC, this does not happen. It is important that we are able to start a graphical session remotely for tasks that cannot be automated such as certain Windows updates.

Prior to uploading this CL, I uninstalled TightVNC on all Windows machines, ran this playbook on all of them, and manually verified that I was able to start a VNC session on each machine.

Bug: skia:12528
Change-Id: Ib97dd9ae32fe4cfef4ad4326b5bd702e43f0cc28
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/517379
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>