Add step to swarming_needs ansible script for gdm3 login.
This checks for the existence of the /etc/gdm3 directory,
and if it exists it writes an updated /etc/gdm3/daemon.conf file
that automatically logs in chrome-bot. This is required
so that on Debian 11 systems, which use Gnome by default,
that test_machine_monitor and all it's child processes can get
access to the X Windows and thus the GPU.

Also adds `swarming_needs__can_build` flag which adds
all the packages needs for a compile of Skia to complete
successfully.

Change-Id: I4b4e80a50bb1ce5b59702d58868f38178313e9bd
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/521940
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>