Delete Infra-PerCommit-{CreateDockerImage,PushAppsFromInfraDockerImage}.
Task Infra-PerCommit-PushAppsFromInfraDockerImage broke when we added support for Go 1.18. This task is used to autodeploy CT and Leasing. After speaking with @rmistry, we decided to delete it because we don't make changes to CT or Leasing very often, and because we're going to migrate to a new scheme to autodeploy all our apps soon. In the meantime, we can push CT and Leasing changes manually.

Task Infra-PerCommit-CreateDockerImage can be deleted because it's used to create the gcr.io/skia-public/infra container, which predates our Bazel build, and is currently only used to build CT and Leasing without Bazel.

Bug: skia:13063
Change-Id: I9ff255d738bb48f48215a35b800a4740e44a83bf
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/523819
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>