[bazel] Make Debugger, particles, et all not require Docker auth when running locally
This also adds the ability to change the port Debugger runs on
via an environment variable, e.g.

DEBUGGER_LOCAL_PORT=:8123 make run-with-custom

The way this is currently handled is a bit awkward due to the fact
that extract_files_from_skia_wasm_container is a macro and thus
cannot actually inspect the flag values (or select statements)
because macros are evaluated before flags are parsed [1].

What we do (an alternative would be replace the macro with a real
rule, but that would take much longer) is load and empty
container if a provided flag is false and touch the output files
to ensure they exist.

[1] https://bazel.build/docs/configurable-attributes#faq-select-macro

Change-Id: Idd2268e7c74527f3684cec6700190d4b6a879d58
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/527216
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>