[gazelle] Add support for absolute labels
This updates the version of gazelle we use to 0.25.0
In particularl, I wanted [1] which adds the ability
to recognize labels like @spirv_tools instead of
having to specify @spirv_tools//:spirv_tools, only
to have that simplified later, as if by magic.

This also replaces our handcrafted label parsing
with the more robust label.Parse provided by gazelle.

This CL is needed to land [2]

[1] https://github.com/bazelbuild/bazel-gazelle/commit/0e5dc95b1ac5ac199f90530831be3fa9022ebfd7
[2] https://skia-review.googlesource.com/c/skia/+/528044

Change-Id: I5ff134aafe5fd15700e2f6526baa8c0af81ba11a
Bug: skia:12541
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/528697
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>