[npm-audit-mirror] Remove maxage:10y
The intention of adding maxage:10y was that the downloaded tarball would
never be updated. Verdaccio's caching is already setup to do that without maxage.

maxage:10y was instead causing the downloaded package metadata to never
be updated. This caused new package versions to not be detected.

Bug: skia:13268
Change-Id: I0a3e38227bb31f3d5f9497ff9278e9d66873fa59
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/534562
Commit-Queue: Ravi Mistry <rmistry@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>