Drop our only used of `replace` in go.mod.
This is the first CL in a two step change.

I have created a branch in
https://github.com/skia-dev/protoc-gen-twirp_typescript/
that changes the name of the module from

   go.larrymyers.com/protoc-gen-twirp_typescript

to

    github.com/skia-dev/protoc-gen-twirp_typescript/

This CL then references that branch in go.mod directly
and drops the `replace` directive. This was done by running
go get and specifying the hash of the commit on the branch.

    $ go get github.com/skia-dev/protoc-gen-twirp_typescript@1ceb8af78d51aaa6cb3f01da848a0a47daa0811f

All other references to go.larrymyers.com/protoc-gen-twirp_typescript
in code are also changed.

After this CL lands then the branch on github.com/skia-dev
can be landed and another CL will be needed to point
go.mod back to github.com/skia-dev/protoc-gen-twirp_typescript@latest.


Change-Id: Ifbf85bf42a3a91029a519e7be15dada8cc497791
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/535196
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>