[bazel] Add helpers under //bazel/cipd to get paths to CIPD binaries.
This CL establishes a pattern to be followed with all CIPD packages: in order to access a CIPD binary, call the corresponding mycipdpackage.FindXXX() function to retrieve its path. By doing so, a Bazel data dependency is automatically added to the corresponding CIPD package, which guarantees that Bazel will download the CIPD package as needed, and no further action is required at the call site.

Bug: skia:13293
Change-Id: Id72b9829e8f9752d95a88cc86ce13679cd54c054
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/542620
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>