[shaders] Pass correct amount of predefined uniforms to children.
We only ever want to pass the 11 (or whatever the current number
of uniforms is). Thus the only meaningful change to
getPredefinedUniformValuesFromControls() to always
return the right number of floats.

The rest of the changes are to clean up some types
and add a few docs that I had to (re)discover while
fixing.

WORKSPACE change is to get https://skia-review.googlesource.com/c/skia/+/543416
when running locally.

Change-Id: I893c28fb3670ef00a721e3f1001b694bf80026e7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/547740
Reviewed-by: Brian Osman <brianosman@google.com>