[npm-audit-mirror] Periodically check creation time of downloaded packages
Packages are downloaded on the mirror only after they pass all security checks (eg: publish age must be greater than a week).
The framework will now, once a day, examine these downloaded packages by querying the global NPM registry and see if their creation time is less than a week. This will indicate that that package was deleted and republished recently, which could happen by a malicious agent.


If a package with creation time of less than a week is found then:
* We check in the DB to see if there is already a corresponding monorail issue filed.
* If no monorail issue exists then we file a new issue and add it to the DB.
* If a monorail issue exists then we check to see if the issue has been closed.
* If the issue has been closed for more than a week then we file a new issue and it to the DB.
* If the issue is still open we do nothing.

The above is done to avoid filing new issues everytime the pod restarts.


Description of code changes:
* Added a new GetDownloadedPackageNames method to the ProjectMirror interface. It examines the local cache and returns a slice of package names.
* Added a new examiner golang package. Once a day it will go through all downloaded packages and will file bugs (as described above) if any packages with creation time < 1 week are found.


Bug: skia:13397
Change-Id: Ia4ffa5c980281475ac7c3ff81c754dd179eeb057
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/548258
Reviewed-by: Joe Gregorio <jcgregorio@google.com>