Update the color pallete of the treemap and the tooltip design
It is generally reccomended to not use the color red to be more color-
blind friendly. In addition, css for the tooltip had been added for to
make it easier to modify. A more distinct outline has been to the
tooltip so that it is not as depenedent on the cell background color
contrast with colors chosen from the colors recommended for tooltips
from material.

Change-Id: Iffbd5e4c418b4bd59501e85484cd537b08acbda0
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/562817
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>
Commit-Queue: Anjuli Jones <anjulij@google.com>