bloaty_treemap: Stop ignoring section data w/o paths
This was originally a heuristic to avoid counting debug symbols and
sections. The CodeSize bots are now excluding that data the right way
(by running bloaty against a stripped copy of the binary).

Ultimately, this heuristic was far too aggressive - there is a
nontrivial amount of data that slips through the cracks (in particular,
the 10 MB embedded ICU data file).

This will effectively add a handful of top-level nodes (one per
section), with a bunch of symbols in each one. We could think about
better ways to visualize this data, but as long as bloaty can't
attribute it to a translation unit, there's only so much we can do.

Bug: skia:12151
Bug: skia:13657
Change-Id: I307e3cb3ab1aba52459b725d2eb65e2cc8f7c513
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/570700
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>