[machine] Implement CPU interrogation for Linux, based on Swarming's.
Move Mac's CPUs() to crossplatform package, as it's useful for Linux
too.

Rename CPUs() smoketest to run on all platforms, as it works on both
Linux and Mac.

Split out ISAAndBitness(), which needed to be separate at one point
during the refactoring but will probably again be useful for Windows.

Change-Id: I8bb3224d101fb6c7b203d0bb98d42742e51d7aa0
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/572278
Commit-Queue: Erik Rose <erikrose@google.com>
Reviewed-by: Eric Boren <borenet@google.com>