[infra] Have Go presubmit use downloaded hermetic buildifier
Instead of building buildifier from scratch (which can take
90+ seconds with a cold cache), we download a prebuilt one.

Originally, I wanted to follow the model for cloud_sdk [1], but
I realized the logic there to download the right binary
for the given system was easier to implement with a select
statement in //bazel/external/buildifier/BUILD.bazel.

This should only download the binary for the target platform,
which is not an issue as long as execution = target platform
(which for the infra repo, should be the case).

Rather than using --run_under, we need to start off in the
Bazel-specified directory so we can find the binaries that
are added to our runfiles (e.g. find the buildifier that is
linked in via deps). The command to run presubmit is now
bazel run //cmd/presubmit -- --repo_dir=$PWD

[1] https://github.com/google/skia-buildbot/blob/20ab6832d47fc3ff42545a57fd3938301b8d63c1/bazel/external/google_cloud_sdk.bzl

Change-Id: I40941516fa9090f855c7f50a7e6914818d66f99d
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/573674
Reviewed-by: Ravi Mistry <rmistry@google.com>