[auth-proxy] Add Role support.
The current flags --cria_group and --allowed_from
are mapped to the Viewer role so auth-proxy remains backward
compatible with current usage.

Also the X-ROLES header is now sent.

This is just the server side. Client side update will be in a
separate CL.

Bug: b/249507110
Change-Id: Id270f4d18a4de82a6032bce97d9060b4e424b9b7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/585905
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>