[bazel] //Makefile: Use --config=mayberemote in "gazelle" and "buildifier" targets.
This CL is similar to https://skia-review.googlesource.com/c/buildbot/+/586699.

Starting from a warm Bazel cache on my workstation, this CL reduced the time it takes to execute the following sequence of commands from 19.91 to 9.6 seconds:

    $ bazel build --config=remote //...
    $ make gazelle
    $ bazel build --config=remote //...

And it reduced the time it takes to execute the following sequence of commands from 17.68 to 5.71 seconds:

    $ bazel build --config=remote //...
    $ make buildifier
    $ bazel build --config=remote //...

I measured the above times in a fashion similar to https://skia-review.googlesource.com/c/buildbot/+/586699. For details, please see that CL's description.

Bug: skia:13785
Change-Id: I04330bfcc918e5675b2687df75dffdaa4ba36dff
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/586700
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>