[bazel] Gazelle extension: Add support for ignoring certain TypeScript imports via "// gazelle:ignore" comments.
This CL makes the following "make gazelle" warnings go away:

    gazelle: Could not find any rules that satisfy import "debugger-app/wasm_libs/types/canvaskit" from @skia_infra//debugger-app/modules/debugger-page-sk (sk_element)
    gazelle: Could not find any rules that satisfy import "debugger-app/wasm_libs/types/canvaskit" from @skia_infra//debugger-app/modules:debugger_ts_lib (ts_library)
    gazelle: Could not find any rules that satisfy import "shaders/wasm_libs/types/canvaskit" from @skia_infra//shaders/modules/shadernode:index_test (karma_test)
    gazelle: Could not find any rules that satisfy import "shaders/wasm_libs/types/canvaskit" from @skia_infra//shaders/modules/shadernode:index_ts_lib (ts_library)
    gazelle: Could not find any rules that satisfy import "shaders/wasm_libs/types/canvaskit" from @skia_infra//shaders/modules/shaders-app-sk (sk_element)

The cause of these warnings is that there are TypeScript imports that refer to files that do not exist in the Git repository but that are generated by Bazel under the //_bazel_bin directory under compilation. Such files are not visible to Gazelle. The solution proposed by this CL is to add a mechanism to tell Gazelle to ignore such imports. For some examples, see the TypeScript files modified by this CL.

Change-Id: Ia708d97042c684582d914815a07ab58c8cb52ab9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/589660
Auto-Submit: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>