[rbe] Add test for failing digest, use consistent ordering in Merge()
I wasn't able to come up with a constructed test case that reproduces
the failure. This is one of the smaller digests which failed during the
outage. I confirmed that this fails with the old implementation and
succeeds with the new.

While I was doing this I noticed that running Merge on the same digest
twice didn't produce the same result digest, because the computation of
the digest depends on the ordering of the slices on the directory nodes.
Adding some sorting fixes the issue.

Change-Id: Id881bc4353d18f07b8c19ee1c655dfc8bd9e0dc1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/591105
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Eric Boren <borenet@google.com>