[machine] Bring Foundry Bot up and down in accordance with machine availability.
Bring it up once we've confirmed machineserver thinks it should be up
(but not until). Take it down promptly when machineserver says it should
be down. If it's in the middle of a task, send it a SIGINT to let it
complete gracefully first.

* Add recentschannel package to facilitate prioritizing the most recent
  machineserver poll result.
* Wire foundrybotcustodian into availability changes by means of a
  new callback.

Change-Id: I2e6841ce3f3f7460316df56ddbeb5163b3a1b599
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/605539
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Erik Rose <erikrose@google.com>