[machine] Interrogating iOS device requires getting the OS Version.
Since the OS version ends up as part of the Dimensions for the machine,
like DeviceType, a failure to interrogate it can't be ignored.

This was found by the presence of Prometheus metrics that were
identical except for os, which in one metric was os="iOS" and
in the other was os="iOS-13.3.1".

Change-Id: I7592a2b7db90b8b46b4b4890400d6fe4f08bd750
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/609117
Reviewed-by: Erik Rose <erikrose@google.com>
Auto-Submit: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>