Remove dependency link emulators -> cockroachdb
While diagnosing slow local builds using Bazel's --profile flag
and Perfetto, I noticed a long time spent on runfile generation.
Looking at some of the generated runfile directories, I noticed
lots of unit tests somehow "depended" on Python, google_cloud_sdk,
and CockroachDb. The first two of these have many files bundled
in, which is probably causing the overhead of sandboxing (i.e.
creating a lot of symlinks or read-only copies).

The reason this is happening is because the emulators package
has a dependency on all three of these things as it is currently
designed as a "one-stop-shop" for all emulators. This means all
executables (mostly unit tests) that indirectly depend on emulators
need to pay extra overhead.

To stop this, I plan to split up the emulators into subpackages
and keep the shared code and constants in emulators. This will
also slim down the testutils/unittest package by moving that
emulator logic to the "proper" emulators package.

emulator "constants" and such will go in the emulators.go file,
start.go will contain the shared logic for starting the emulators
(which will be called by the subpackages). For now, start.go
has copies of logic found elsewhere, but those copies will go
away after migration.

Change-Id: Id9a308c8358c7805bb984c6f318e63ae24f8ab4f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/612297
Reviewed-by: Joe Gregorio <jcgregorio@google.com>