Build all taskdrivers on the CI
This also adds building the linux taskdrivers to the CI.

In https://skia-review.googlesource.com/c/buildbot/+/600519
support for using pre-built task drivers was added. As a
result, we no longer build the task drivers at every
commit, which could lead to silent breakages.

This remedies that. Note these tasks would be run in
parallel to the jobs that do not require them, keeping
the reduced latency of the linked CL.

Change-Id: Ib46147ca73347096d4a84e791a3838d72370b9b7
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/614497
Reviewed-by: Eric Boren <borenet@google.com>