Add Skylab service account to Chrome instance
Adds the Skylab service account to the list of service accounts that
are allowed to upload to the Chrome Gold instance. We are in the process
of enabling GPU tests in Skylab, and one of the remaining issues is that
our pixel tests do not have permission to upload to Gold from there.

Bug: chromium:1399908
Change-Id: I343dd1af1898a96a843c4b4d6c0553a1f496e0af
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/616852
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>