[skolo] Install Skia's Foundry Bot variant on Linux x86-64.
It is part of the install_test_machine_monitor role because it has a few
tight couplings with the installation of TMM. Most prominently, it needs
to upload a renamed version of Foundry Bot on Windows (because running
executables can't be overwritten) and then have the startup job drop it
into place later. It saves a lot of trouble to have this be the same
startup job that does this to the TMM executable. If FB had its own job,
we'd have to ensure the FB one ran first, lest TMM start up quickly and
launch the old FB.

Bug: b/261753362
Change-Id: Ia6eb03ac18f238921f6e80cde76003e660152571
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/629057
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Erik Rose <erikrose@google.com>