Move quarantine message and metrics from Processor to store.Update.
Move the following code, which was in Processor.Process() to
store.Update:

    _ = machine.SetSwarmingQuarantinedMessage(&newD)
    SetQuarantineMetrics(newD)

That's because this code doesn't require any information
from machine.Event, so it doesn't need to be in Processor.Process(),
and store.Update() is a better location for enforcing such
global validations and corrections on machine.Descriptions.

Also a drive-by cleanup removing Description.HasValidPool()
and machine.SetSwarmingPool(), which should have been removed
in the CL that added pool configuration.

Change-Id: I7ed7873be9551e4b1684415e1525b548af07dfc2
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/634837
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>