Add sample_area inexact matching
Adds the sample_area inexact matching algorithm. This is meant for
large/complex images that are prone to noise. Both images are compared
in small, user-defined sample areas and are considered the same image
if less than a certain number of pixels per sample area differ. Also
supports setting a tolerance so that barely differing pixels are
treated as identical.

Bug: angleproject:7985
Change-Id: I7acb286e4e0883732ad52696b627a3ae591d663e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/636030
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>