fiddle: Improve GCS write error handling
Replace deferred calls to `util.Close(w)` with explicit calls
to `w.Close()`. This is needed because it is possible, when
the data being written is smaller than the write buffer, for
a `w.Write(...)` to succeed, but the close to fail - which is
when the HTTP post is made.

Change-Id: I235b09493b94ed570e19637992cf272d61cd5c38
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/652420
Commit-Queue: Chris Mumford <cmumford@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>