scrap: Add make variables to simplify port conflict resolution
Fiddle makes calls to scrap exchange. Add make variables to simplify
the specification of a few ports to avoid a conflict when both
fiddle and scrap exchange want to listen on the sama port.

This change allows scrapexchange, fiddle, and jsfiddle to all
be started on the same host without the need to modify any
makefiles to avoid port conflicts.

Change-Id: I84352f8745f75c833a2f029e7218930626275d8e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/654656
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>