[gold] /json/v2/search RPC: Exclude digests with optional key disallow_triaging=true from BulkTriageDeltaInfos.
For CLs, this change adds the overhead of looking up the optional keys for each extendedBulkTriageDeltaInfo. I don't expect this to add significant latency to the RPC because it uses the options cache[1], and because the query for cache misses should be fast.

For the primary branch, we have the additional step of fetching the optionIDs for each digest. This is done via a single query against the TraceValues table (inspired in [2]). I expect this query to be pretty fast because the WHERE clause leverages the table's index.

[1] https://skia.googlesource.com/buildbot/+/264f3ee077677f5eddf47808a3f33987322f05a6/golden/go/search/search.go#1627
[2] https://skia.googlesource.com/buildbot/+/264f3ee077677f5eddf47808a3f33987322f05a6/golden/go/search/search.go#1818

Bug: skia:14033
Change-Id: I55765b64308e826b949e87e06ecc9b5fc3297d42
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/658056
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>