[goldctl] goldclient_test.go: Add more image matching algorithm tests.
This CL adds some missing tests for the recently added sample_area image matching algorithm.

Additionally, this CL adds tests for all algorithms for the case where there is no existing positive image. This is important because in a follow-up CL I will define a new algorithm, positive_if_only_image. We need to call this algorithm with a nil reference image when there is no existing positive image. However, the way image matching algorithms currently work is that we don't call them at all if there is no reference image, and fail early instead. Rather than treating positive_if_only_image as a special case, my follow-up CL will change the flow so that all algorithms receive a nil reference image if there is no existing positive image (with the exception of the exact algorithm, which we continue to treat as a special case - all we do is check whether or not the input image is already labeled as positive).

Bug: skia:14033
Change-Id: I7ee310287be88c109308b131bc5046e5d4eed54a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/658257
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Auto-Submit: Leandro Lovisolo <lovisolo@google.com>