[perf] Factor out the git/provider.Provider interface.
This is an initial step in adding Gitiles support, as it factors
out the only touchpoints with git, makes that into an interface (provider.Provider),
and then moves all the git specific code into an implementation
of that interface (providers/cli).

A follow-on CL will implement providers.Provider for Gitiles.

This also adds a provider selection to the config file format
and updates all the existing configs to use 'git'.

Finally this adds an optional StartCommit, as opposed to always
starting from the initial commit to the repo, which would be
prohibitive in a huge repo like Chrome.

Bug: b/271555267
Change-Id: Ie8bb28656684532cf91b39e933d98890435b9b0c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/651947
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>