[perf] Change gitiles provider to use LogFnBatch.
On first startup of a new repo CommitsFromMostRecentGitHashToHead
query for a full set of logs for the entire repo, which is huge
for large repos like Chrome, so switch to LogFnBatch and also
make sure the query repsponse is in reverse.

Finally add a Makefile target for testing these changes locally.

Change-Id: I508b818fec6d0becae788ea0df98d522f4c30e95
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/666897
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>