Add TLS endpoint to envoy config.
This adds a second endpoint to the envoy config for `skia-infra-public` only.

It does that by duplicating the `listener` entry, changing the port it listens
on, adds in the TLS config, and then appends it back as a new `listener`.

The resulting file was validated using envoy. The only error is because
I tested this on my desktop and not in the container we build, so the
cert files are missing.

```
$ docker run -ti -v $(pwd):/tmp/configs envoyproxy/envoy:v1.16.1  --mode validate --config-path /tmp/configs/skia-infra-public.json
[2023-04-07 20:51:23.222][8][info][main] [source/server/server.cc:583] runtime: layers:
  - name: base
    static_layer:
      {}
  - name: admin
    admin_layer:
      {}
[2023-04-07 20:51:23.222][8][info][config] [source/server/configuration_impl.cc:95] loading tracing configuration
[2023-04-07 20:51:23.223][8][info][config] [source/server/configuration_impl.cc:70] loading 0 static secret(s)
[2023-04-07 20:51:23.223][8][info][config] [source/server/configuration_impl.cc:76] loading 30 cluster(s)
[2023-04-07 20:51:23.229][8][info][config] [source/server/configuration_impl.cc:80] loading 2 listener(s)
[2023-04-07 20:51:23.236][8][critical][main] [source/server/config_validation/server.cc:60] error initializing configuration '/tmp/configs/skia-infra-public.json': Invalid path: /etc/envoy/front-proxy-crt.pem
```

See: https://www.envoyproxy.io/docs/envoy/latest/start/quick-start/securing#upstream-and-downstream-tls-contexts
Change-Id: I3d650984b4694e62ff3c5ab34150820d349ff157
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/667459
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>