Remove login and logout URL from proxylogin and alogin-sk.
Instead make the simplifying assumption that login and logout will be done at
the domain (skia.org, or luci.app), as opposed to the sub-domain the application
is running at.

In contrast, requesting the login status, which includes the user's roles in the
response, must go to the sub-domain.

Bug: b/279736504
Change-Id: I74083a3b52738a9f86a535cccc4837e0abb10791
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/681599
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>