[k8s-config] Add byte ranges to ParseK8sConfigFile
This helps when we're editing k8s config files and want something in
between fully re-encoding everything via the YAML encoder (which would
drop all formatting and comments) and performing a simple
find-and-replace.

Note - gopkg.in/yaml.v3 provides an intermediate format in yaml.Node
which does preserve comments (but not whitespace, though Node does
contain line and column information, so it may be possible to
reconstruct to some extent). It is not, however, straightforward to
decode the YAML onto a struct while maintaining a mapping back to the
yaml.Node so that we could re-encode the YAML and preserve comments. It
would be a fairly involved process to do this at anything deeper than
the document level. This is complicated by the fact that the Kubernetes
API uses a custom YAML decoding process which uses JSON as an
intermediate format (to allow reuse of [Un]MarshalJSON and `json:""`
struct tags.

Bug: skia:14284
Change-Id: Icec22444a5694d28e960f0ba02674d12f753ef1f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/690836
Commit-Queue: Eric Boren <borenet@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>