Issue Fix: Ingestor Error: Failed to find git hash for commit number: -1
The ingestor process had a bug, which uses types.BadCommitNumber as the
input to get the relevent git hash from Commits table.
The consiquence of that bug is: git.GetCommitNumber will fail after the
git.Update(), which will fail the ingestion of that GCS file.

Change-Id: I2268877892bc157a4819eff44cc69585edb2fd91
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/693730
Commit-Queue: Alex Sun <sunpeng@google.com>
Reviewed-by: Funing Wang <funing@google.com>