Fix race condition in fiddle.
Writing the images to GCS was previously done in a Go routine
which meant that sometimes the image URLs would be returned
before the images made it to GCS. Then GCS would return a 404.

All the writes are still done in Go routines, but we use a WaitGroup to ensure
that all images are written before returning the results.

Change-Id: I3200a36f41a5d21fcfa705db559f91a7f027d94e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/696876
Commit-Queue: Ravi Mistry <rmistry@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>
Auto-Submit: Joe Gregorio <jcgregorio@google.com>