Add option to only ACK a message after the file was parsed successfully.
When there is traffic spike (for example, backfill bunch of files to GCS bucket), there will be some GCS client timeout error: "net/http: request canceled (Client.Timeout or context cancellation while reading body)".
We have already enabled the GCS http client failure retry, but it is still hard to know if a timeout failure is eventually resolved by the retry or not.

The current message handling logic is:
1. Pull a message from pub/sub.
2. Validate and handle the message,
  a. if there is non retry-able issue (e.g. cannot load message content, file name is not under configured GCS bucket), ACK the message and continue to handle the next message.
  b. if there is retry-able issue, NACK the message. the message will be pulled by an ingestor again.
  c. if there is no issue, ACK the message, and put the file into the chan, which will be handled by the ingestor worker.
3. The ingestor worker read file from GCS, parse the file into paramset and traces, and write the data into DB tables.

This CL is to add another option (based on configuration file) to the pub/sub message ACK behavior, the new message ACK behavior will be:
Do not ACK message until the file is parsed and the trace values are written into DB successfully.
If a message was never ACKed after exceed max delivery attempts, it will be moved to dead letter topic.
We need to investigate the message in dead letter topic, then we may disregard it or replay it if necessary.

To address this CL, we need to make two configuration changes to the pub/sub configurations:
1. Increase the Acknowledgment deadline from default value (10 seconds ) to 10 minutes, that will help to make sure during the ingestor worker read/parse file and write data into DB, the same message won't be available to another ingestor worker. Pub/Sub doc here: https://cloud.google.com/pubsub/docs/create-subscription#subscription_properties
2. Setup dead letter topic for the pub/sub, thus a message will be put to the dead letter topic automatically once it reached to the max delivery attempts (default is 5).
3. Setup metrics & alarm on the size of dead letter topic.

Change-Id: I0660a40e6f84a6f1a64d6554c874e364de71a4a9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/701616
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Alex Sun <sunpeng@google.com>